Split protocol version tests into dedicated module
